requirements.txt
# Web Framework
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.9
jinja2>=3.1.3

# Web Automation
selenium>=4.18.1
webdriver-manager>=4.0.1

# PowerPoint Generation
python-pptx>=0.6.23
Pillow>=10.2.0

# Testing
pytest>=8.0.2
pytest-cov>=4.1.0

# Linting and Formatting
black>=24.2.0
pylint>=3.0.3
flake8>=7.0.0
mypy>=1.8.0

# Type Checking
types-requests>=2.25.0
types-python-dateutil>=2.8.2

# Development Tools
python-dotenv>=0.19.0 